fix : move collection files in folder
