+++ src/content/config.ts
@@ -1,13 +1,12 @@
 import { z } from "astro:content";
-import { talksCollection } from "./talks";
-import { eventsCollection } from "./events";
-import { organizersCollection } from "./organizers";
+import { talksCollection } from "./talks/talks";
+import { eventsCollection } from "./events/events";
+import { organizersCollection } from "./organizers/organizers";
 import { partnersCollection } from "./partners";
-import { speakersCollection } from "./speaker";
-import { sponsorsCollection } from "./sponsors";
+import { speakersCollection } from "./speakers/speaker";
+import { sponsorsCollection } from "./sponsors/sponsors";
 
-export const zSocialTypes = () =>
-  z.enum(["x", "linkedin", "instagram", "github"]);
+export const zSocialTypes = z.enum(["x", "linkedin", "instagram", "github"]);
 
 export const collections = {
   organizers: organizersCollection,
